Changed some internal structures from string to char* and kstring_t
